requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
brotli==1.2.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
dj-database-url==0.5.0
django==4.2.26
django-cors-headers==4.4.0
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
Pillow==10.1.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
whitenoise==6.7.0
